refactor(runtime): migrate sc-batch-integration preflight to omicsclaw/runtime/preflight/sc_batch.py — closes #118
Fifth slice of the bot/core decomposition (PRD #113, ADR 0001). Moves
~390 LOC of single-cell domain business logic out of the entry-layer
``bot/core.py`` into a new ``omicsclaw/runtime/preflight/`` package
sitting alongside the other Skill-aware runtime modules
(``system_prompt``, ``tool_registry``, ``predicates``).

The accident of placement — sc-batch-integration's batch-key detection
ended up in bot/core because the agent loop happened to need it first
— is now corrected. Future Skills that need preflight checks land their
own modules under ``omicsclaw.runtime.preflight``; the package's
``__init__.py`` documents this growth path.

New files:

* ``omicsclaw/runtime/preflight/__init__.py`` — placeholder noting the
  intent to grow into a generic preflight engine that consults a
  Skill's declared prerequisite schema. No exported API.
* ``omicsclaw/runtime/preflight/sc_batch.py`` — 14 functions migrated
  unchanged in behaviour:
  - 3 scoring tables (``_BATCH_KEY_EXACT_PREFERENCES``,
    ``_BATCH_KEY_HINT_TERMS``, ``_BATCH_KEY_EXCLUDED_COLUMNS``)
  - Pure helpers: ``_normalize_obs_key``, ``_extract_flag_value``,
    ``_resolve_requested_batch_key``
  - AnnData-touching: ``_load_h5ad_obs_dataframe``,
    ``_score_batch_key_candidate``, ``_find_batch_key_candidates``,
    ``_inspect_h5ad_integration_readiness``
  - Renderers: ``_format_batch_key_clarification``,
    ``_format_sc_batch_workflow_guidance``,
    ``_format_auto_prepare_summary``
  - Workflow: ``_maybe_require_batch_key_selection``,
    ``_get_sc_batch_integration_workflow_plan``,
    ``_maybe_require_batch_integration_workflow``,
    ``_auto_prepare_sc_batch_integration``

Cross-module concerns (late imports inside functions):

* ``_lookup_skill_info`` is still in ``bot.core`` (will move to
  ``bot.skill_orchestration`` in #119); ``_maybe_require_batch_key
  _selection`` and ``_get_sc_batch_integration_workflow_plan`` late-
  import it.
* ``_run_omics_skill_step`` and ``execute_omicsclaw`` are still in
  ``bot.core`` (will move to ``bot.skill_orchestration`` and
  ``bot.tool_executors``); ``_auto_prepare_sc_batch_integration``
  late-imports them.

Backward-compat: ``bot/core.py`` re-exports every previously-public
symbol so ``bot.core._maybe_require_batch_key_selection``,
``bot.core._auto_prepare_sc_batch_integration``, etc. all resolve
through the new module.

Tests added (``tests/runtime/preflight/test_sc_batch.py``, 9 cases):

* Pure helpers: ``_normalize_obs_key`` punctuation collapsing;
  ``_resolve_requested_batch_key`` precedence direct > extra_args.
* Scoring: rejects excluded columns (cluster / annotation / QC);
  high score for exact preference match (``sample``); rejects
  singleton or per-cell-unique columns; preview includes example
  values.
* Clarification rendering: lists candidates when found; explains
  when a requested key is not in obs columns.

Verification:

* ``tests/runtime/preflight/`` — 9 / 9 pass.
* Bot regression net (``tests/bot/``, ``test_bot_*``,
  ``test_oauth_regressions``, ``test_skill_listing``,
  ``test_routing_regression``, ``test_auto_routing_disambiguation``,
  ``test_app_server``, ``test_notebook_files``, ``test_bot_core_timeout``):
  225 passed, 1 skipped, 4 xfailed — no regressions.
* ``tests/eval/`` + ``tests/test_knowhow.py``: 121 passed, 17
  deselected. Eval baseline 14/15 against ``deepseek-v4-pro``
  unaffected; KH-de-padj-guardrails surfacing logic untouched.
* ``bot/core.py``: 4318 → 3880 LOC (−438). Cumulative drop from main:
  5160 → 3880 (−1280, −24.8%).
* ``omicsclaw/runtime/preflight/sc_batch.py`` lands at ~440 LOC.

Closes #118.
